Lection 2

+++ src/main/java/com/gorvic/gshop/dto/ProductDto.java
@@ -4,13 +4,15 @@ import com.gorvic.gshop.models.Product;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 public class ProductDto {
     private Long id;
     private String title;
     private String categoryTitle;
-    private float price;
+    private BigDecimal price;
 
     public ProductDto(Product product) {
         this.id = product.getId();
@@ -19,7 +21,7 @@ public class ProductDto {
         this.price = product.getPrice();
     }
 
-    public ProductDto(Long id, String title, String categoryTitle, float price) {
+    public ProductDto(Long id, String title, String categoryTitle, BigDecimal price) {
         this.id = id;
         this.title = title;
         this.categoryTitle = categoryTitle;
